src/requirements.txt
boto3==1.34.108
botocore==1.34.108
confluent-kafka==2.4.0
jmespath==1.0.1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
